+++ tests/models/element/test_attributes.py
@@ -1,0 +1,37 @@
+from tests.utils import TestElement
+
+
+def test_kwargs_stored_in_attributes() -> None:
+    element = TestElement(test="green")
+
+    assert element.attributes["test"] == "green"
+
+
+def test_render_attributes() -> None:
+    element = TestElement(test="green")
+
+    assert element._render_attributes() == " test='green'"
+
+
+def test_render_multiple_attributes_separated_by_space() -> None:
+    element = TestElement(one="one", two="two")
+
+    assert element._render_attributes() == " one='one' two='two'"
+
+
+def test_false_value_is_skipped() -> None:
+    element = TestElement(test=False)
+
+    assert element._render_attributes() == ""
+
+
+def test_true_value_is_added_as_key_only() -> None:
+    element = TestElement(test=True)
+
+    assert element._render_attributes() == " test"
+
+
+def test_htmx_keys_added_with_hyphen() -> None:
+    element = TestElement(hx_get="url")
+
+    assert element._render_attributes() == " hx-get='url'"
