+++ tests/utils.py
@@ -1,0 +1,6 @@
+from hypermedia.models import Element
+
+
+class TestElement(Element):
+    def dump(self) -> str:
+        return str(self)
